+++ Apps/hoppy-hub/src/app/auth/login-modal/login-modal.component.ts
@@ -16,7 +16,10 @@ import {
 } from '@angular/forms';
 import { ModalService, ModalType } from '../../services/modal.service';
 import { Router } from '@angular/router';
-import { AlertService, AlertType } from '../../alert/alert.service';
+import {
+  AlertService,
+  AlertType
+} from '../../shared-components/alert/alert.service';
 import { AuthService } from '../auth.service';
 
 @Component({
