+++ Apps/hoppy-hub/src/app/auth/register-modal/register-modal.component.ts
@@ -18,7 +18,10 @@ import { ModalService, ModalType } from '../../services/modal.service';
 import { Router } from '@angular/router';
 import { CustomValidators } from '../../shared/custom-validators';
 import { AuthService } from '../auth.service';
-import { AlertService, AlertType } from '../../alert/alert.service';
+import {
+  AlertService,
+  AlertType
+} from '../../shared-components/alert/alert.service';
 
 @Component({
   selector: 'app-register-modal',
