+++ Apps/hoppy-hub/src/app/recent-opinions/recent-opinions.component.ts
@@ -5,7 +5,7 @@ import { Subscription } from 'rxjs';
 import { Opinion } from '../opinions/opinion.model';
 import { PagedList } from '../shared/paged-list';
 import { OpinionsParams } from '../opinions/opinions-params';
-import { LoadingSpinnerComponent } from '../loading-spinner/loading-spinner.component';
+import { LoadingSpinnerComponent } from '../shared-components/loading-spinner/loading-spinner.component';
 import { ErrorMessageComponent } from '../shared-components/error-message/error-message.component';
 
 @Component({
